Move all callback examples into one file
