+++ hydra/experimental/callbacks.py
@@ -7,14 +7,30 @@ from typing import Any
 
 from omegaconf import DictConfig
 
-from hydra.core.utils import JobReturn
+from hydra.core.utils import JobReturn, JobStatus
 from hydra.experimental.callback import Callback
 
-log = logging.getLogger(__name__)
+
+class LogJobReturnCallback(Callback):
+    def __init__(self) -> None:
+        self.log = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+
+    def on_job_end(
+        self, config: DictConfig, job_return: JobReturn, **kwargs: Any
+    ) -> None:
+        if job_return.status == JobStatus.COMPLETED:
+            self.log.info(f"Succeeded with return value: {job_return.return_value}")
+        elif job_return.status == JobStatus.FAILED:
+            self.log.error("", exc_info=job_return._return_value)
+        else:
+            self.log.error("Status unknown. This should never happen.")
 
 
 class PickleJobInfoCallback(Callback):
     output_dir: Path
+
+    def __init__(self) -> None:
+        self.log = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
 
     def on_job_start(self, config: DictConfig, **kwargs: Any) -> None:
         self.output_dir = Path(config.hydra.runtime.output_dir) / Path(
@@ -22,14 +38,14 @@ class PickleJobInfoCallback(Callback):
         )
         filename = "config.pickle"
         self._save_pickle(obj=config, filename=filename, output_dir=self.output_dir)
-        log.info(f"Saving job configs in {self.output_dir / filename}")
+        self.log.info(f"Saving job configs in {self.output_dir / filename}")
 
     def on_job_end(
         self, config: DictConfig, job_return: JobReturn, **kwargs: Any
     ) -> None:
         filename = "job_return.pickle"
         self._save_pickle(obj=job_return, filename=filename, output_dir=self.output_dir)
-        log.info(f"Saving job_return in {self.output_dir / filename}")
+        self.log.info(f"Saving job_return in {self.output_dir / filename}")
 
     def _save_pickle(self, obj: Any, filename: str, output_dir: Path) -> None:
         output_dir.mkdir(parents=True, exist_ok=True)
