+++ tests/test_apps/app_with_pickle_job_info_callback/config.yaml
@@ -3,4 +3,4 @@ foo: bar
 hydra:
   callbacks:
     save_job_info:
-      _target_: hydra.experimental.pickle_job_info_callback.PickleJobInfoCallback
+      _target_: hydra.experimental.callbacks.PickleJobInfoCallback
