+++ website/docs/experimental/callbacks.md
@@ -5,6 +5,9 @@ sidebar_label: Callbacks
 ---
 
 import GithubLink from "@site/src/components/GithubLink"
+import {ExampleGithubLink} from "@site/src/components/GithubLink"
+
+<ExampleGithubLink text="Examples" to="hydra/experimental/callbacks.py"/>
 
 The <GithubLink to="hydra/experimental/callback.py">Callback interface</GithubLink> enables custom
 code to be triggered by various Hydra events.
